adding upload order scenario as caseworker with the new judge title/name fields
