+++ e2e/fixtures/manageCases/caseTabs/C100/c100DraftOrdersTabContent.ts
@@ -1,6 +1,5 @@
 export enum C100DraftOrdersTabContent {
-  tabName = "Draft orders",
   tabTitle = "Draft orders",
   subTitle1 = "Oder made by",
-  typeOfOrder = "Child arrangements, specific issue or prohibited steps order (C43)",
+  typeOfOrderField = "Type of order",
 }
