+++ e2e/fixtures/manageCases/caseWorker/createAnOrder/orderCA/manageOrders5CAContent.ts
@@ -23,13 +23,12 @@ export enum ManageOrders5CAContent {
   label18 = "Full name of Justices' Legal Adviser (Optional)",
   label19 = "Date order made",
   label20 = "Is the order about all the children?",
-  label21 = "Add recitals or preamble (Optional)",
-  label22 = "Add directions (Optional)",
-  hiddenLabel1 = "At which hearing was the order approved?",
+  labelCreate1 = "Add recitals or preamble (Optional)",
+  labelCreate2 = "Add directions (Optional)",
+  hiddenLabel1 = "Was the order approved at a hearing?",
   hiddenLabel2 = "Day",
   hiddenLabel3 = "Month",
   hiddenLabel4 = "Year",
   hearing = "No hearings available",
   judgeFullName = "judgeFirstName judgeLastName",
-  approvedAtAHearing = "Was the order approved at a hearing?",
 }
