+++ e2e/pages/manageCases/caseTabs/C100/c100DraftOrdersTabPage.ts
@@ -1,18 +1,15 @@
-import { Page } from "@playwright/test";
+import { Page, expect } from "@playwright/test";
 import { Selectors } from "../../../../common/selectors";
-import { Helpers } from "../../../../common/helpers";
 import AccessibilityTestHelper from "../../../../common/accessibilityTestHelper";
 import { C100DraftOrdersTabContent } from "../../../../fixtures/manageCases/caseTabs/C100/c100DraftOrdersTabContent.ts";
+import { ManageOrders5CAContent } from "../../../../fixtures/manageCases/caseWorker/createAnOrder/orderCA/manageOrders5CAContent.ts";
 
 export class C100DraftOrdersTabPage {
   public static async c100DraftOrdersTabPage(
     page: Page,
     accessibilityTest: boolean,
   ): Promise<void> {
-    await this.checkPageLoads(
-      page,
-      accessibilityTest,
-    );
+    await this.checkPageLoads(page, accessibilityTest);
   }
 
   private static async checkPageLoads(
@@ -22,15 +19,15 @@ export class C100DraftOrdersTabPage {
     await page.waitForSelector(
       `${Selectors.h1}:text-is("${C100DraftOrdersTabContent.tabTitle}")`,
     );
-    await Promise.all([
-      await Helpers.checkGroup(
-        page,
-        4,
-        C100DraftOrdersTabContent,
-        "subTitle",
-        `${Selectors.h2}`,
-      ),
-    ]);
+    await expect(
+      page.getByText(C100DraftOrdersTabContent.typeOfOrderField),
+    ).toBeVisible();
+    await expect(
+      page.getByText(C100DraftOrdersTabContent.subTitle1),
+    ).toBeVisible();
+    await expect(
+      page.getByText(ManageOrders5CAContent.judgeFullName),
+    ).toBeVisible();
     if (accessibilityTest) {
       await AccessibilityTestHelper.run(page);
     }
